@@ -3,25 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
-    "react-scripts": "3.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-router-bootstrap": "^0.25.0",
+    "@fortawesome/fontawesome-free": "^5.11.2",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-regular-svg-icons": "^5.11.2",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@fortawesome/react-fontawesome": "^0.1.7",
     "bootstrap3": "^3.3.5",
-    "react-bootstrap": "0.32.4",
     "query-string": "^6.8.3",
+    "react": "^16.12.0",
+    "react-bootstrap": "^0.32.4",
     "react-cookie": "^4.0.1",
+    "react-dom": "^16.12.0",
     "react-keydown": "^1.9.12",
-    "react-redux": "^7.1.1",
+    "react-redux": "^7.1.3",
+    "react-router-bootstrap": "^0.25.0",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "3.3.0",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.25",
-    "@fortawesome/free-solid-svg-icons": "^5.11.2",
-    "@fortawesome/free-regular-svg-icons": "^5.11.2",
-    "@fortawesome/fontawesome-free": "^5.11.2",
-    "@fortawesome/react-fontawesome": "^0.1.5"
+    "webpack": "4.41.2",
+    "core-js": "^3.6.0",
+    "typescript": "^3.7.3",
+    "whatwg-fetch": "^3.0.0"
   },
   "scripts": {
     "start": "react-scripts start --host localhost --port 3000",
